refactor: fix preview screen in android studio
