+++ app/src/main/kotlin/com/jeeldobariya/passcodes/navigation/NavigationRoot.kt
@@ -22,12 +22,12 @@ import com.jeeldobariya.passcodes.password_manager.presentation.save_password.Mo
 import com.jeeldobariya.passcodes.password_manager.presentation.update_password.ClassicalUpdatePasswordScreen
 import com.jeeldobariya.passcodes.password_manager.presentation.update_password.ModernUpdatePasswordScreen
 import com.jeeldobariya.passcodes.password_manager.presentation.view_password.ClassicalViewPasswordScreen
-import com.jeeldobariya.passcodes.ui.ClassicalAboutScreen
-import com.jeeldobariya.passcodes.ui.ClassicalMainScreen
-import com.jeeldobariya.passcodes.ui.ClassicalSettingsScreen
-import com.jeeldobariya.passcodes.ui.ModernAboutScreen
-import com.jeeldobariya.passcodes.ui.ModernMainScreen
-import com.jeeldobariya.passcodes.ui.ModernSettingsScreen
+import com.jeeldobariya.passcodes.presentation.about_screen.ClassicalAboutScreen
+import com.jeeldobariya.passcodes.presentation.main_screen.ClassicalMainScreen
+import com.jeeldobariya.passcodes.presentation.setting_screen.ClassicalSettingsScreen
+import com.jeeldobariya.passcodes.presentation.about_screen.ModernAboutScreen
+import com.jeeldobariya.passcodes.presentation.main_screen.ModernMainScreen
+import com.jeeldobariya.passcodes.presentation.setting_screen.ModernSettingsScreen
 
 
 @Composable
